视频模块-视频删除

+++ service/service-video/src/main/java/com/linyilinyi/video/controller/VideoController.java
@@ -64,7 +64,7 @@ public class VideoController {
 
     @Operation(summary = "删除视频")
     @DeleteMapping("/deleteVideo/{ids}")
-    public Result<String> deleteVideo(@NotNull(message = "ids不能为空") @PathVariable List<Long> ids) {
+    public Result<String> deleteVideo(@NotNull(message = "ids不能为空") @PathVariable List<Integer> ids) {
         return Result.ok(videoService.deleteVideo(ids));
     }
