+++ service/service-video/src/main/java/com/linyilinyi/video/service/VideoService.java
@@ -27,7 +27,7 @@ public interface VideoService extends IService<Video> {
 
     Video addVideo(VideoAddVo video);
 
-    String deleteVideo(List<Long> ids);
+    String deleteVideo(List<Integer> ids);
 
     String updateVideo(Video video);
